// src/api/api.schemas.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Messaging API
 * OpenAPI spec version: 1.0.0
 */
export interface User {
  id: number;
  username: string;
}

export interface ThreadId {
  id: number;
}

/**
 * @nullable
 */
export type ThreadSummaryLastMessage = {
  id?: number;
  content?: string;
  createdAt?: string;
} | null;

export interface ThreadSummary {
  id: number;
  createdAt: string;
  participants: User[];
  /** @nullable */
  lastMessage?: ThreadSummaryLastMessage;
}

export interface Message {
  id: number;
  threadId: number;
  createdAt: string;
  content: string;
  author: User;
}

export type LoginBody = {
  username: string;
  password: string;
};

export type Login200 = {
  ok?: boolean;
  username?: string;
};

export type CreateThreadBody = {
  username: string;
};

export type ListMessagesParams = {
sinceId?: number;
limit?: number;
};

export type ListMessages200 = {
  items?: Message[];
  /** @nullable */
  nextCursor?: number | null;
};

export type SendMessageBody = {
  /** @minLength 1 */
  content: string;
};